Extend the API to allow one to have overlapping samples (async)
Experiments to allow one to start up a range operation and not have the processor locked up while waiting for the range to complete.  So there is a start operation, functions to query to see if complete (or maybe check IO pin state) and then function to get the results and clear the interrupt state.